requirements.txt
streamlit
pinecone
google-generativeai
cohere
python-dotenv